+++ src/PSRule/Commands/NewRuleDefinitionCommand.cs
@@ -70,6 +70,17 @@ namespace PSRule.Commands
 
             context.VerboseFoundRule(ruleName: Name, scriptName: MyInvocation.ScriptName);
 
+            var visitor = new RuleLanguageAst(ruleName: Name, context: context);
+            Body.Ast.Visit(visitor);
+
+            if (visitor.Errors != null)
+            {
+                foreach (var errorRecord in visitor.Errors)
+                {
+                    WriteError(errorRecord: errorRecord);
+                }
+            }
+
             var ps = PowerShell.Create();
             ps.Runspace = context.GetRunspace();
             ps.AddCommand(new CmdletInfo(InvokeBlockCmdletName, typeof(InvokeRuleBlockCommand)));