+++ tests/PSRule.Tests/FromFileNested.Rule.ps1
@@ -1,0 +1,10 @@
+#
+# Pester unit test rules checking nesting
+#
+
+# Synopsis: Should error with nested rule
+Rule 'WithNestedRule' {
+    Rule 'NestedRule' {
+        $True;
+    }
+}